requirements.txt
pytest==6.2.5
Django==4.0
django-cors-headers==3.10.1
django-rest-framework==0.1.0
djangorestframework==3.13.0
psycopg2-binary==2.9.2